DEV-1310 make test coverage 100%

+++ wr.go
@@ -1,21 +1,19 @@
 package main
 
 import (
-	"crypto/md5"
 	"database/sql"
-	"encoding/hex"
 	"fmt"
-	"io/ioutil"
+	"io/fs"
 	"os"
 	"os/exec"
 	"path"
 	"path/filepath"
 	"runtime"
-	"sort"
 	"strings"
 	"syscall"
 	"time"
 
+	"github.com/bep/debounce"
 	"github.com/gen2brain/beeep"
 	"github.com/web-ridge/wr/helpers"
 
@@ -32,6 +30,7 @@ var (
 	restart = make(chan bool)
 	port    = os.Getenv("PORT")
 )
+var db *sql.DB
 
 func main() {
 	// let us have colored logs
@@ -63,8 +62,6 @@ func start(c *cli.Context) error {
 	fmt.Println("               _     _____  _     _            \n              | |   |  __ \\(_)   | |           \n __      _____| |__ | |__) |_  __| | __ _  ___ \n \\ \\ /\\ / / _ \\ '_ \\|  _  /| |/ _` |/ _` |/ _ \\\n  \\ V  V /  __/ |_) | | \\ \\| | (_| | (_| |  __/\n   \\_/\\_/ \\___|_.__/|_|  \\_\\_|\\__,_|\\__, |\\___|\n                                     __/ |     \n                                    |___/   ") //nolint:lll
 	fmt.Println("")
 
-	notify("test", "test")
-
 	backendPath, err := os.Getwd()
 	checkError("cant get current dir", err)
 	startPath := filepath.Dir(backendPath)
@@ -80,18 +77,19 @@ func start(c *cli.Context) error {
 
 	// wait till the db is started
 	time.Sleep(1 * time.Second)
-	db := helpers.WaitForDatabase()
+	db = helpers.WaitForDatabase()
 
-	dropDatabase(db)
+	dropDatabase()
 	runMigrations()
 
 	runSeeder()
 	runConvertPlugin()
 
 	// start watching migrations/code
-	go watch(db, backendPath, frontendPath)
+	go watch(backendPath, frontendPath)
 
 	// start server and wait for restarts
+	killPortProcess(port)
 	existingServer := startServerInBackground(false)
 	for <-restart {
 		log.Debug().Msg("restarting backend...")
@@ -107,21 +105,22 @@ func start(c *cli.Context) error {
 }
 
 func notify(title, message string) {
-	err := beeep.Notify(title, message, "assets/information.png")
+	err := beeep.Notify(title, message, "./icon.png")
 	checkError("could not notify", err)
 }
 
 func stopServer(existingServer *exec.Cmd) {
 	// https://stackoverflow.com/a/68179972/2508481
 	// Send kill signal to the process group instead of single process (it gets the same value as the PID, only negative)
-	err := syscall.Kill(-existingServer.Process.Pid, syscall.SIGKILL)
-	checkError("can not stop server", err)
+	if existingServer != nil && existingServer.Process != nil {
+		err := syscall.Kill(-existingServer.Process.Pid, syscall.SIGKILL)
+		checkError("can not stop server", err)
+	}
 
 	killPortProcess(port)
 }
 
 func startServerInBackground(restart bool) *exec.Cmd {
-	killPortProcess(port)
 	cmd := exec.Command("/bin/sh", "-c", fmt.Sprintf("WR_RESTART=%v go run server.go", restart))
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
@@ -134,7 +133,11 @@ func startServerInBackground(restart bool) *exec.Cmd {
 		checkServerError(err)
 		defer func() {
 			log.Debug().Msg("kill server")
-			checkError("can not kill server", cmd.Process.Kill())
+			err = cmd.Process.Kill()
+			alreadyKilled := strings.Contains(err.Error(), "process already finished")
+			if !alreadyKilled {
+				checkError("can not kill server", err)
+			}
 		}()
 	}()
 
@@ -157,7 +160,7 @@ func runMigrations() {
 	return
 }
 
-func dropDatabase(db *sql.DB) {
+func dropDatabase() {
 	name := os.Getenv("DATABASE_NAME")
 	var err error
 	_, err = db.Exec(fmt.Sprintf("DROP DATABASE IF EXISTS `%v`", name))
@@ -206,7 +209,53 @@ func execKillCommand(cmd *exec.Cmd) {
 	}
 }
 
-func runRelayGenerator(backendPath string) {
+func runRelayGenerator() {
+	runMergeSchemas()
+	runRelay()
+	forceIndexGeneratedDirectory()
+}
+
+// forceIndexGeneratedDirectory forces the frontend IDE to index the generated dir faster
+func forceIndexGeneratedDirectory() {
+	prefix := "wr-version-index-"
+	files := glob("../frontend/src/__generated__", func(s string) bool {
+		return strings.HasPrefix(s, prefix)
+	})
+	for _, file := range files {
+		log.Debug().Str("file", file).Msg("loop version index")
+		checkError("could not force index of relay.dev (existing)", os.Rename(file, "./wr-version-index-"))
+	}
+	if len(files) == 0 {
+		f, err := os.Create(prefix + "0")
+		checkError("could not close file", f.Close())
+		checkError("could not force index of relay.dev (new)", err)
+	}
+}
+
+func runMergeSchemas() {
+	log.Debug().Msg("run merge-schemas")
+
+	cmd := exec.Command("yarn", "merge-schemas")
+	cmd.Dir = "../frontend"
+	// cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	err := cmd.Run()
+	checkError("failed to run convert/convert.go", err)
+
+	log.Debug().Msg("✅ done with merge-schemas")
+}
+
+func runRelay() {
+	log.Debug().Msg("run relay.dev")
+
+	cmd := exec.Command("yarn", "relay")
+	cmd.Dir = "../frontend"
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	err := cmd.Run()
+	checkError("failed to run convert/convert.go", err)
+
+	log.Debug().Msg("✅ done with relay.dev")
 }
 
 func runSeeder() {
@@ -224,25 +273,7 @@ func runSeeder() {
 	log.Debug().Msg("✅ done seed/seed.go")
 }
 
-func haveMigrationsChanged() bool {
-	migrationsPath := path.Join("./")
-	migrationsHashPath := path.Join(migrationsPath, ".wr")
-	nextMigrationHash, err := MD5AllString(migrationsPath)
-	checkError("error getting hash of migrations", err)
-
-	// TODO: get hash of all migration contents
-	currentMigrationsHash, err := os.ReadFile(migrationsHashPath)
-
-	changed := string(currentMigrationsHash) != nextMigrationHash
-	if changed {
-		// TODO: run migration here + write new hash to file
-		err = os.WriteFile(migrationsHashPath, []byte(nextMigrationHash), 0o644)
-		checkError("error writing hash of migrations", err)
-	}
-	return changed
-}
-
-func watch(db *sql.DB, backendPath, frontendPath string) {
+func watch(backendPath, frontendPath string) {
 	watcher, err := fsnotify.NewWatcher()
 	if err != nil {
 		log.Fatal().Err(err).Msg("can not start file watcher")
@@ -264,7 +295,7 @@ func watch(db *sql.DB, backendPath, frontendPath string) {
 				}
 				if event.Op&fsnotify.Write == fsnotify.Write {
 					log.Debug().Str("file", event.Name).Msg("modified file")
-					fileChanged(db, event)
+					fileChanged(event)
 				}
 			case err, ok := <-watcher.Errors:
 				if !ok {
@@ -275,12 +306,10 @@ func watch(db *sql.DB, backendPath, frontendPath string) {
 		}
 	}()
 
-	// TODO: watch migrations folder
-	// TODO: watch custom_schema.grapql if changed run convert_plugin
-
 	filesOrDirectoriesToWatch := []string{
-		backendPath,
-		path.Join(frontendPath, "schema_custom.graphql"),
+		backendPath + "/*",
+		"../frontend/schema_custom.graphql",
+		"../frontend/__generated__",
 	}
 	for _, w := range filesOrDirectoriesToWatch {
 		err = watcher.Add(backendPath)
@@ -290,24 +319,67 @@ func watch(db *sql.DB, backendPath, frontendPath string) {
 	<-done
 }
 
-func fileChanged(db *sql.DB, event fsnotify.Event) {
+var debounced = debounce.New(200 * time.Millisecond)
+
+func runSqlChanged() {
+	dropDatabase()
+	runMigrations()
+	runConvertPlugin()
+	runSeeder()
+	restart <- true
+}
+
+func runSchemaChanged() {
+	runConvertPlugin()
+	runRelayGenerator()
+	restart <- true
+}
+
+func runSeedChanged() {
+	runSeeder()
+}
+
+func runGoChanged() {
+	restart <- true
+}
+
+func runGeneratedChanged() {
+	forceIndexGeneratedDirectory()
+}
+
+func runMigrationsChanged() {
+	runMigrations()
+	runConvertPlugin()
+	restart <- true
+}
+
+func fileChanged(event fsnotify.Event) {
+	modelsChanged := strings.Contains(event.Name, "/models/")
 	envChanged := strings.Contains(event.Name, ".env")
 	goChanged := strings.Contains(event.Name, ".go")
 	sqlChanged := strings.Contains(event.Name, ".sql")
 	schemaChanged := strings.Contains(event.Name, ".graphql")
+	seedChanged := strings.Contains(event.Name, "/seed/")
+	generatedChanged := strings.Contains(event.Name, "/__generated__/")
+	migrationsChanged := strings.Contains(event.Name, "/migrations/")
+	// we only change models from here so we don't need to subscribe
+	if modelsChanged {
+		return
+	}
 
 	switch true {
 	case sqlChanged:
-		dropDatabase(db)
-		runMigrations()
-		runConvertPlugin()
-		runSeeder()
-		restart <- true
+		debounced(runSqlChanged)
 	case schemaChanged:
-		runConvertPlugin()
-		restart <- true
+		debounced(runSchemaChanged)
+	case seedChanged:
+		debounced(runSeedChanged)
 	case goChanged, envChanged:
-		restart <- true
+		debounced(runGoChanged)
+	case generatedChanged:
+		debounced(runGeneratedChanged)
+	case migrationsChanged:
+		debounced(runMigrationsChanged)
 	}
 }
 
@@ -321,46 +393,18 @@ func startDbInDocker() *exec.Cmd {
 
 func checkError(s string, err error) {
 	if err != nil {
+		notify("Error 🔥🔥🔥", s)
 		log.Fatal().Err(err).Msg("🔥🔥🔥 " + s)
 	}
 }
 
-func MD5AllString(root string) (string, error) {
-	m, err := MD5All(root)
-	if err != nil {
-		return "", err
-	}
-	var values []string
-	for p, v := range m {
-		k := strings.TrimPrefix(p, root+"/")
-		pass := hex.EncodeToString(v[:])
-		values = append(values, k+"="+pass)
-	}
-	sort.Strings(values)
-	return strings.Join(values, "\n"), nil
-}
-
-// MD5All reads all the files in the file tree rooted at root and returns a map
-// from file path to the MD5 sum of the file's contents.  If the directory walk
-// fails or any read operation fails, MD5All returns an error.
-func MD5All(root string) (map[string][md5.Size]byte, error) {
-	m := make(map[string][md5.Size]byte)
-	err := filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
+func glob(root string, fn func(string) bool) []string {
+	var files []string
+	filepath.WalkDir(root, func(s string, d fs.DirEntry, e error) error {
+		if fn(s) {
+			files = append(files, s)
 		}
-		if !info.Mode().IsRegular() {
-			return nil
-		}
-		data, err := ioutil.ReadFile(path)
-		if err != nil {
-			return err
-		}
-		m[path] = md5.Sum(data)
 		return nil
 	})
-	if err != nil {
-		return nil, err
-	}
-	return m, nil
+	return files
 }
